refactor: 카드 생성 시 개수 유효성 검사 이동
