+++ src/test/java/blackjack/domain/card/CardFactoryTest.java
@@ -18,4 +18,17 @@ public class CardFactoryTest {
         Assertions.assertThat(cards).hasSize(expectedSize);
 
     }
+
+    // TODO : 예외처리 리팩토링 고민해보기
+    @Test
+    @DisplayName("카드개수가 48개가 아니라면 예외처리한다.")
+    void cards_init_validate_test() {
+        // given & when
+        List<Card> cards = CardFactory.of();
+
+        // then
+        Assertions.assertThatThrownBy(() -> Cards.init())
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("카드의 개수는 총 48개여야 합니다.");
+    }
 }
